refactor keypad system
